+++ NbuildTasks/DownloadFile.cs
@@ -1,4 +1,3 @@
-
 using Ntools;
 using System;
 using System.IO;